requirements.txt
pillow
aiohttp
bcrypt
beautifulsoup4
colorama
discord.py
fade
fake-useragent
instaloader
mysql-connector-python
pandas
phonenumbers
pyfiglet
pynput
pypresence
pystyle
requests
selenium
undetected-chromedriver
webdriver-manager
tls-client
